requirements.txt
django==3.0.3
django-tinymce4-lite
django-bootstrap4==1.0.1
django-crispy-forms==1.8.1